refactor(analyzer): separate comb loop procedure and regions
