+++ crates/analyzer/src/comb_loop_detect.rs
@@ -8,7 +8,10 @@
 //! boxes, `inout` ports, recursive functions) add no edges; the
 //! simulator's `analyze_dependency` is the backup safety net.
 
-mod ssa;
+mod procedure;
+mod region;
+
+use region::{ArraySpan, BitPartition, IdxKey, NodeKey, PackedSpan, dst_writes, var_reads};
 
 use crate::AnalyzerError;
 use crate::HashMap;
@@ -17,7 +20,7 @@ use crate::conv::Context;
 use crate::ir::VarId;
 use crate::ir::{
     AssignDestination, Component, Declaration, Expression, Factor, FunctionCall, InstDeclaration,
-    Ir, Module, Op, Statement, SystemFunctionKind, VarIndex, VarSelect, Variable,
+    Ir, Module, Op, Statement, SystemFunctionKind, VarSelect, Variable,
 };
 use crate::symbol::{Affiliation, Direction};
 use daggy::petgraph::Graph;
@@ -26,117 +29,6 @@ use daggy::petgraph::graph::NodeIndex;
 use daggy::petgraph::visit::EdgeRef;
 use std::collections::VecDeque;
 use veryl_parser::resource_table::StrId;
-
-#[derive(Clone, Copy, Debug, Default, PartialEq, Eq, Hash, PartialOrd, Ord)]
-struct ArraySpan {
-    start: usize,
-    length: usize,
-}
-
-impl ArraySpan {
-    fn end(self) -> Option<usize> {
-        self.start.checked_add(self.length)
-    }
-
-    fn overlaps(self, other: Self) -> bool {
-        let Some(left_end) = self.end() else {
-            return false;
-        };
-        let Some(right_end) = other.end() else {
-            return false;
-        };
-        self.start < right_end && other.start < left_end
-    }
-}
-
-/// Half-open packed-bit interval. Unlike a `BigUint` mask, its storage and
-/// operations are independent of the declared bit width.
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, PartialOrd, Ord)]
-struct PackedSpan {
-    start: usize,
-    length: usize,
-}
-
-impl PackedSpan {
-    fn new(start: usize, length: usize) -> Option<Self> {
-        (length != 0 && start.checked_add(length).is_some()).then_some(Self { start, length })
-    }
-
-    fn whole(width: usize) -> Option<Self> {
-        Self::new(0, width)
-    }
-
-    fn from_select(high: usize, low: usize) -> Option<Self> {
-        Self::new(low, high.checked_sub(low)?.checked_add(1)?)
-    }
-
-    fn end(self) -> usize {
-        self.start + self.length
-    }
-
-    fn overlaps(self, other: Self) -> bool {
-        self.start < other.end() && other.start < self.end()
-    }
-
-    fn intersection(self, other: Self) -> Option<Self> {
-        let start = self.start.max(other.start);
-        let end = self.end().min(other.end());
-        Self::new(start, end.checked_sub(start)?)
-    }
-
-    fn translated(self, from: usize, to: usize) -> Option<Self> {
-        let start = self.start.checked_sub(from)?.checked_add(to)?;
-        Self::new(start, self.length)
-    }
-}
-
-/// One split unpacked-array interval. Bit precision lives in packed spans.
-type IdxKey = (VarId, ArraySpan);
-
-/// `(VarId, array_idx, range_idx)`. `range_idx` indexes the variable's
-/// `BitPartition`, so bit-disjoint reads/writes form disjoint nodes.
-type NodeKey = (VarId, ArraySpan, usize);
-
-/// Per `IdxKey`, atomic packed-bit intervals.
-#[derive(Default)]
-struct BitPartition {
-    ranges: HashMap<IdxKey, Vec<PackedSpan>>,
-}
-
-impl BitPartition {
-    /// Empty slice means the variable's bits are untouched.
-    fn ranges_of(&self, key: IdxKey) -> &[PackedSpan] {
-        self.ranges.get(&key).map(|v| v.as_slice()).unwrap_or(&[])
-    }
-
-    fn overlapping(&self, key: IdxKey, span: PackedSpan) -> Vec<usize> {
-        self.ranges_of(key)
-            .iter()
-            .enumerate()
-            .filter(|(_, range)| range.overlaps(span))
-            .map(|(i, _)| i)
-            .collect()
-    }
-
-    fn overlapping_access(&self, id: VarId, access: ArraySpan, span: PackedSpan) -> Vec<NodeKey> {
-        let mut keys = self
-            .ranges
-            .iter()
-            .filter(|((object, _), _)| *object == id)
-            .filter(|((_, split), _)| split.overlaps(access))
-            .flat_map(|((_, split), ranges)| {
-                ranges
-                    .iter()
-                    .enumerate()
-                    .filter(|(_, range)| range.overlaps(span))
-                    .map(|(range, _)| (id, *split, range))
-            })
-            .collect::<Vec<_>>();
-        keys.sort_unstable();
-        keys.dedup();
-        keys
-    }
-}
 
 /// `feedthrough[child_in_id] = { child_out_ids reachable purely combinationally }`.
 /// Port-level only -- the parent keeps bit precision via `BitPartition`.
@@ -999,7 +891,7 @@ fn build_module_graph(
         let Declaration::Comb(comb) = declaration else {
             continue;
         };
-        for (source, destination) in ssa::analyze(module, &bit_part, &comb.statements) {
+        for (source, destination) in procedure::analyze(module, &bit_part, &comb.statements) {
             if !is_module_scope_var(source.0, &module.variables)
                 || !is_module_scope_var(destination.0, &module.variables)
             {
@@ -1405,100 +1297,6 @@ fn compute_module_summary(module: &Module, graph: &Graph<NodeKey, ()>) -> Module
         }
     }
     ModuleCombSummary { feedthrough }
-}
-
-/// Mirrors the packed-region logic of `AssignDestination::eval_assign`.
-fn dst_writes(dst: &AssignDestination, ctx: &mut Context) -> Vec<(ArraySpan, PackedSpan)> {
-    let Some(variable) = ctx.get_variable_info(dst.id) else {
-        return Vec::new();
-    };
-    let is_select_const = dst.select.is_const();
-
-    let packed = if !is_select_const {
-        let Some(packed) = conservative_select_span(&dst.select, &variable.r#type, ctx) else {
-            return Vec::new();
-        };
-        packed
-    } else {
-        let Some((high, low)) = dst.select.eval_value(ctx, &variable.r#type, false) else {
-            return Vec::new();
-        };
-        let Some(packed) = PackedSpan::from_select(high, low) else {
-            return Vec::new();
-        };
-        packed
-    };
-
-    array_access_span(&dst.index, &variable.r#type, ctx)
-        .map(|span| vec![(span, packed)])
-        .unwrap_or_default()
-}
-
-fn var_reads(
-    id: VarId,
-    index: &VarIndex,
-    select: &VarSelect,
-    ctx: &mut Context,
-) -> Vec<(ArraySpan, PackedSpan)> {
-    let Some(variable) = ctx.variables.get(&id).cloned() else {
-        return Vec::new();
-    };
-    let packed = if select.is_const_with_range()
-        && let Some((high, low)) = select.eval_value(ctx, &variable.r#type, false)
-    {
-        let Some(packed) = PackedSpan::from_select(high, low) else {
-            return Vec::new();
-        };
-        packed
-    } else {
-        let Some(packed) = conservative_select_span(select, &variable.r#type, ctx) else {
-            return Vec::new();
-        };
-        packed
-    };
-    array_access_span(index, &variable.r#type, ctx)
-        .map(|span| vec![(span, packed)])
-        .unwrap_or_default()
-}
-
-fn array_access_span(
-    index: &VarIndex,
-    r#type: &crate::ir::Type,
-    ctx: &mut Context,
-) -> Option<ArraySpan> {
-    let prefix_len = index
-        .0
-        .iter()
-        .take_while(|expression| expression.comptime().is_const)
-        .count();
-    let prefix = VarIndex(index.0[..prefix_len].to_vec());
-    let values = prefix.eval_value(ctx)?;
-    let (start, inclusive_end) = r#type.array.calc_range(&values)?;
-    Some(ArraySpan {
-        start,
-        length: inclusive_end.checked_sub(start)?.checked_add(1)?,
-    })
-}
-
-fn conservative_select_span(
-    select: &VarSelect,
-    r#type: &crate::ir::Type,
-    ctx: &mut Context,
-) -> Option<PackedSpan> {
-    let prefix = VarSelect(
-        select
-            .0
-            .iter()
-            .take_while(|expression| expression.comptime().is_const)
-            .cloned()
-            .collect(),
-        None,
-    );
-    if let Some((high, low)) = prefix.eval_value(ctx, r#type, false) {
-        PackedSpan::from_select(high, low)
-    } else {
-        r#type.total_width().and_then(PackedSpan::whole)
-    }
 }
 
 #[cfg(test)]
